1.0.4 remove version pin on json-2.1.0
in version 2.2.0 the json gem introduced a dependency which requires
build_essentials in order to compile native extensions. this landed
as a breaking change until we figured out why our gem installs were
failing so we temporarily pinned the json gem to 2.1.0 which did
not require build_essentials.

if you are unable to install build_essentials you may be able to
use simple_approvals-chefspec 1.0.3 which should be functional
with a version pin on json-2.1.0

Now that this dependency issue is understood we should no longer
need this pin.